Refactor ChunkManager
Splits our current implementation of ChunkManager into an abstract base class (ChunkManager), an IndexingChunkManager, and a CachingChunkManager.
