+++ kaldb/src/main/java/com/slack/kaldb/chunk/manager/indexing/ChunkRollOverInProgressException.java
@@ -1,4 +1,4 @@
-package com.slack.kaldb.chunk;
+package com.slack.kaldb.chunk.manager.indexing;
 
 public class ChunkRollOverInProgressException extends RuntimeException {
   public ChunkRollOverInProgressException(String message) {
