+++ kaldb/src/main/java/com/slack/kaldb/chunk/manager/indexing/IndexingChunkManager.java
@@ -1,57 +1,45 @@
-package com.slack.kaldb.chunk;
+package com.slack.kaldb.chunk.manager.indexing;
 
 import static com.slack.kaldb.util.ArgValidationUtils.ensureNonNullString;
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
 
-import brave.ScopedSpan;
-import brave.Tracing;
 import com.google.common.annotations.VisibleForTesting;
-import com.google.common.util.concurrent.*;
-import com.linecorp.armeria.common.RequestContext;
+import com.google.common.util.concurrent.FutureCallback;
+import com.google.common.util.concurrent.Futures;
+import com.google.common.util.concurrent.ListenableFuture;
+import com.google.common.util.concurrent.ListeningExecutorService;
+import com.google.common.util.concurrent.MoreExecutors;
 import com.slack.kaldb.blobfs.s3.S3BlobFs;
 import com.slack.kaldb.blobfs.s3.S3BlobFsConfig;
+import com.slack.kaldb.chunk.Chunk;
+import com.slack.kaldb.chunk.ReadWriteChunkImpl;
+import com.slack.kaldb.chunk.RollOverChunkTask;
+import com.slack.kaldb.chunk.SearchContext;
+import com.slack.kaldb.chunk.manager.ChunkManager;
 import com.slack.kaldb.config.KaldbConfig;
 import com.slack.kaldb.logstore.LogMessage;
 import com.slack.kaldb.logstore.LogStore;
 import com.slack.kaldb.logstore.LuceneIndexStoreImpl;
-import com.slack.kaldb.logstore.search.SearchQuery;
-import com.slack.kaldb.logstore.search.SearchResult;
-import com.slack.kaldb.logstore.search.SearchResultAggregator;
-import com.slack.kaldb.logstore.search.SearchResultAggregatorImpl;
 import com.slack.kaldb.metadata.search.SearchMetadata;
 import com.slack.kaldb.metadata.search.SearchMetadataStore;
 import com.slack.kaldb.metadata.snapshot.SnapshotMetadataStore;
 import com.slack.kaldb.proto.config.KaldbConfigs;
 import com.slack.kaldb.server.MetadataStoreService;
-import com.spotify.futures.CompletableFutures;
 import io.micrometer.core.instrument.MeterRegistry;
 import java.io.File;
 import java.io.IOException;
 import java.time.Instant;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.*;
+import java.util.concurrent.SynchronousQueue;
+import java.util.concurrent.ThreadPoolExecutor;
+import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicLong;
-import java.util.stream.Collectors;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-/**
- * The log data in each Kaldb indexer is stored as chunks. Each chunk is backed by a single instance
- * of log store. Each log store instance which internally contains a lucene index. So, a chunk is
- * identical to a shard.
- *
- * <p>Chunk manager provides a unified api to write and query all the chunks in the application. The
- * addMessage api is used by writers where as the query API is used by the readers.
- *
- * <p>Internally the chunk manager maintains a list of chunks. All chunks except one is considered
- * active. The chunk manager writes the message to the currently active chunk. Once a chunk reaches
- * a roll over point(defined by a roll over strategy), the current chunk is marked as read only. At
- * that point a new chunk is created which becomes the active chunk.
- */
-public class ChunkManager<T> extends AbstractIdleService {
-  private static final Logger LOG = LoggerFactory.getLogger(ChunkManager.class);
+public class IndexingChunkManager<T> extends ChunkManager<T> {
+  private static final Logger LOG = LoggerFactory.getLogger(IndexingChunkManager.class);
   public static final long DEFAULT_ROLLOVER_FUTURE_TIMEOUT_MS = 30000;
 
   private final File dataDirectory;
@@ -59,13 +47,6 @@ public class ChunkManager<T> extends AbstractIdleService {
   // TODO: ChunkDataPrefix can be moved to KaldbConfig?
   private final String chunkDataPrefix;
 
-  // ConcurrentHashMap here guarantees we will not have race conditions when updating and reading
-  // the chunkMap. There is a possibility when reading that the values will be stale, so care
-  // should be taken to ensure that the possibility of the chunk having already been removed is
-  // considered.
-  private final Map<String, Chunk<T>> chunkMap = new ConcurrentHashMap<>();
-
-  // TODO: Pass a reference to BlobFS instead of S3BlobFS.
   private final S3BlobFs s3BlobFs;
   private final String s3Bucket;
   private final ChunkRollOverStrategy chunkRollOverStrategy;
@@ -85,15 +66,8 @@ public class ChunkManager<T> extends AbstractIdleService {
   private final long rolloverFutureTimeoutMs;
   private ListenableFuture<Boolean> rolloverFuture;
 
-  // TODO: We want to move this to the config eventually
-  // Less than KaldbDistributedQueryService#READ_TIMEOUT_MS
-  public static final int QUERY_TIMEOUT_SECONDS = 10;
-  public static final int LOCAL_QUERY_THREAD_POOL_SIZE = 4;
-
   // TODO: Pass this in via config file.
   private static final String CHUNK_DATA_PREFIX = "log_data_";
-
-  private static final ExecutorService queryExecutorService = queryThreadPool();
 
   /**
    * A flag to indicate that ingestion should be stopped. Currently, we only stop ingestion when a
@@ -126,8 +100,7 @@ public class ChunkManager<T> extends AbstractIdleService {
         MoreExecutors.getExitingExecutorService(rollOverExecutor));
   }
 
-  // TODO: Pass in BlobFs object instead of S3BlobFS to make this more generic.
-  public ChunkManager(
+  public IndexingChunkManager(
       String chunkDataPrefix,
       String dataDirectory,
       ChunkRollOverStrategy chunkRollOverStrategy,
@@ -163,15 +136,6 @@ public class ChunkManager<T> extends AbstractIdleService {
         "Created a chunk manager with prefix {} and dataDirectory {}",
         chunkDataPrefix,
         dataDirectory);
-  }
-
-  /*
-     One day we will have to think about rate limiting/backpressure and we will revisit this so it could potentially reject threads if the pool is full
-  */
-  private static ExecutorService queryThreadPool() {
-    return Executors.newFixedThreadPool(
-        LOCAL_QUERY_THREAD_POOL_SIZE,
-        new ThreadFactoryBuilder().setNameFormat("chunk-manager-query-%d").build());
   }
 
   /**
@@ -294,105 +258,6 @@ public class ChunkManager<T> extends AbstractIdleService {
     return activeChunk;
   }
 
-  /*
-   * Query the chunks in the time range, aggregate the results per aggregation policy and return the results.
-   * We aggregate locally and and then the query aggregator will aggregate again. This is OKAY for the current use-case we support
-   * 1. topK results sorted by timestamp
-   * 2. histogram over a fixed time range
-   * We will not aggregate locally for future use-cases that have complex group by etc
-   */
-  public CompletableFuture<SearchResult<T>> query(SearchQuery query) {
-
-    SearchResult<T> errorResult =
-        new SearchResult<>(new ArrayList<>(), 0, 0, new ArrayList<>(), 0, 0, 1, 0);
-
-    List<CompletableFuture<SearchResult<T>>> queries =
-        chunkMap
-            .values()
-            .stream()
-            .filter(
-                chunk ->
-                    chunk.containsDataInTimeRange(
-                        query.startTimeEpochMs / 1000, query.endTimeEpochMs / 1000))
-            .map(
-                (chunk) ->
-                    CompletableFuture.supplyAsync(
-                            () -> {
-                              ScopedSpan span =
-                                  Tracing.currentTracer()
-                                      .startScopedSpan("ReadWriteChunkImpl.query");
-                              span.tag("chunkId", chunk.info().chunkId);
-                              span.tag(
-                                  "chunkCreationTimeSecsSinceEpoch",
-                                  String.valueOf(chunk.info().getChunkCreationTimeEpochSecs()));
-                              span.tag(
-                                  "chunkLastUpdatedTimeSecsEpochSecs",
-                                  String.valueOf(
-                                      chunk.info().getChunkLastUpdatedTimeSecsEpochSecs()));
-                              span.tag(
-                                  "dataStartTimeEpochSecs",
-                                  String.valueOf(chunk.info().getDataStartTimeEpochSecs()));
-                              span.tag(
-                                  "dataEndTimeEpochSecs",
-                                  String.valueOf(chunk.info().getDataEndTimeEpochSecs()));
-                              span.tag(
-                                  "chunkSnapshotTimeEpochSecs",
-                                  String.valueOf(chunk.info().getChunkSnapshotTimeEpochSecs()));
-                              span.tag("numDocs", String.valueOf(chunk.info().getNumDocs()));
-                              span.tag("chunkSize", String.valueOf(chunk.info().getChunkSize()));
-                              span.tag("readOnly", String.valueOf(chunk.isReadOnly()));
-
-                              try {
-                                return chunk.query(query);
-                              } finally {
-                                span.finish();
-                              }
-                            },
-                            RequestContext.makeContextPropagating(queryExecutorService))
-                        // TODO: this will not cancel lucene query. Use ExitableDirectoryReader in
-                        // the future and pass this timeout
-                        .orTimeout(QUERY_TIMEOUT_SECONDS, TimeUnit.SECONDS))
-            .map(
-                chunkFuture ->
-                    chunkFuture.exceptionally(
-                        err -> {
-                          LOG.warn("Chunk Query Exception " + err);
-                          // We catch IllegalArgumentException ( and any other exception that
-                          // represents a parse failure ) and instead of returning an empty result
-                          // we throw back an error to the user
-                          if (err.getCause() instanceof IllegalArgumentException) {
-                            throw (IllegalArgumentException) err.getCause();
-                          }
-                          return errorResult;
-                        }))
-            .collect(Collectors.<CompletableFuture<SearchResult<T>>>toList());
-
-    // TODO: if all fails return error instead of empty and add test
-
-    // Using the spotify library ( this method is much easier to operate then using
-    // CompletableFuture.allOf and converting the CompletableFuture<Void> to
-    // CompletableFuture<List<SearchResult>>
-    CompletableFuture<List<SearchResult<T>>> searchResults = CompletableFutures.allAsList(queries);
-
-    // Increment the node count right at the end so that we increment it only once
-    //noinspection unchecked
-    return ((SearchResultAggregator<T>) new SearchResultAggregatorImpl<>(query))
-        .aggregate(searchResults)
-        .thenApply(this::incrementNodeCount);
-  }
-
-  private SearchResult<T> incrementNodeCount(SearchResult<T> searchResult) {
-    return new SearchResult<>(
-        searchResult.hits,
-        searchResult.tookMicros,
-        searchResult.totalCount,
-        searchResult.buckets,
-        searchResult.failedNodes,
-        searchResult.totalNodes + 1,
-        searchResult.totalSnapshots,
-        searchResult.snapshotsWithReplicas);
-  }
-
   public void removeStaleChunks(List<Map.Entry<String, Chunk<T>>> staleChunks) {
     if (staleChunks.isEmpty()) return;
 
@@ -430,18 +295,13 @@ public class ChunkManager<T> extends AbstractIdleService {
   }
 
   @VisibleForTesting
-  public Map<String, Chunk<T>> getChunkMap() {
-    return chunkMap;
-  }
-
-  @VisibleForTesting
   public ListenableFuture<?> getRolloverFuture() {
     return rolloverFuture;
   }
 
   @Override
   protected void startUp() throws Exception {
-    LOG.info("Starting chunk manager");
+    LOG.info("Starting indexing chunk manager");
     metadataStoreService.awaitRunning(KaldbConfig.DEFAULT_START_STOP_DURATION);
 
     // TODO: Allow multiple nodes to register against the /search path
@@ -486,7 +346,7 @@ public class ChunkManager<T> extends AbstractIdleService {
    */
   @Override
   protected void shutDown() throws IOException {
-    LOG.info("Closing chunk manager.");
+    LOG.info("Closing indexing chunk manager.");
 
     // Stop executor service from taking on new tasks.
     rolloverExecutorService.shutdown();
@@ -518,26 +378,26 @@ public class ChunkManager<T> extends AbstractIdleService {
     }
     // searchMetadataStore.close();
     // snapshotMetadataStore.close();
-    LOG.info("Closed chunk manager.");
+    LOG.info("Closed indexing chunk manager.");
   }
 
-  public static ChunkManager<LogMessage> fromConfig(
+  public static IndexingChunkManager<LogMessage> fromConfig(
       MeterRegistry meterRegistry,
       MetadataStoreService metadataStoreService,
       KaldbConfigs.ServerConfig serverConfig) {
     ChunkRollOverStrategy chunkRollOverStrategy = ChunkRollOverStrategyImpl.fromConfig();
 
     // TODO: Read the config values for chunk manager from config file.
-    ChunkManager<LogMessage> chunkManager =
-        new ChunkManager<>(
+    IndexingChunkManager<LogMessage> chunkManager =
+        new IndexingChunkManager<>(
             CHUNK_DATA_PREFIX,
             KaldbConfig.get().getIndexerConfig().getDataDirectory(),
             chunkRollOverStrategy,
             meterRegistry,
             getS3BlobFsClient(KaldbConfig.get()),
             KaldbConfig.get().getS3Config().getS3Bucket(),
-            ChunkManager.makeDefaultRollOverExecutor(),
-            ChunkManager.DEFAULT_ROLLOVER_FUTURE_TIMEOUT_MS,
+            makeDefaultRollOverExecutor(),
+            DEFAULT_ROLLOVER_FUTURE_TIMEOUT_MS,
             metadataStoreService,
             SearchContext.fromConfig(serverConfig));
